+++ bindings/python/iota_wallet/account.py
@@ -2,15 +2,17 @@ from iota_wallet.common import send_message_routine
 
 
 class Account:
-    def __init__(self, alias_index, handle):
-        self.alias_index = alias_index
+    def __init__(self, account_id, handle):
+        self.account_id = account_id
         self.handle = handle
 
     @send_message_routine
     def __str__(self):
         message = {
             'cmd': 'getAccount',
-            'payload': self.alias_index,
+            'payload': {
+                'accountId': self.account_id,
+            }
         }
         return message
 
@@ -19,7 +21,7 @@ class Account:
         message = {
             'cmd': 'callAccountMethod',
             'payload': {
-                'accountId': self.alias_index,
+                'accountId': self.account_id,
                 'method': {
                     'name': method,
                 }